Even out partition assignment
In a situation where there are more consumers than partitions on all
topics, some consumers will be assigned no work at all with the current
code.

For example, with 2 topics, each with 2 partitions, and 3
consumers running, each topic's partition 0 will be assigned to the
first member ID, and partition 1 will be assigned to the second member
ID. This means 2 members with 2 partitions to consume, and 1 with
nothing!

With this change, all topic:partitions are considered together. In the
above example, 2 members would get 1 partition each, and one would be
assigned 2 partitions.